+++ src/main/java/me/pajic/affogato_core/iris/ShaderPresetSetup.java
@@ -2,21 +2,32 @@ package me.pajic.affogato_core.iris;
 
 import me.pajic.affogato_core.ClientMain;
 import net.fabricmc.loader.api.FabricLoader;
+import net.fabricmc.loader.api.SemanticVersion;
+import net.fabricmc.loader.api.Version;
+import net.fabricmc.loader.api.VersionParsingException;
 import org.apache.commons.io.FileUtils;
 import org.jetbrains.annotations.Nullable;
 
+import java.io.BufferedWriter;
 import java.io.File;
+import java.io.FileWriter;
 import java.io.IOException;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.util.Arrays;
 import java.util.Objects;
+import java.util.Optional;
+import java.util.Set;
+import java.util.stream.Collectors;
 import java.util.stream.Stream;
 
+// sets up Euphoria Patches presets for use in Affogato when shader config files are provided in a correct format
+// inside the `shaderpacks/affogato_presets` directory and a Euphoria Patches shader file is present
 public class ShaderPresetSetup {
 
     private static final Path SHADER_DIR = FabricLoader.getInstance().getGameDir().resolve("shaderpacks");
     private static final Path SHADER_PRESETS_DIR = SHADER_DIR.resolve("affogato_presets");
+    private static final Path EP_VERSION_PATH = SHADER_DIR.resolve("affogato_ep_ver");
     public static @Nullable String[] presets = null;
     public static boolean allPresetsCreated = false;
 
@@ -24,7 +35,9 @@ public class ShaderPresetSetup {
     public static void init() {
         File shaderDir = SHADER_DIR.toFile();
         File shaderPresetsDir = SHADER_PRESETS_DIR.toFile();
+        SemanticVersion epVersion = null;
         if (shaderDir.exists() && shaderPresetsDir.exists() && shaderDir.isDirectory() && shaderPresetsDir.isDirectory()) {
+            // check if the preset files are correctly named and add them
             presets = new String[shaderPresetsDir.listFiles().length];
             try (Stream<Path> fileStream = Files.walk(SHADER_PRESETS_DIR, 1)) {
                 fileStream.filter(path -> !path.toFile().getName().equals("affogato_presets"))
@@ -40,39 +53,78 @@ public class ShaderPresetSetup {
                 ClientMain.LOGGER.error("Can't access file, did not create shader presets", e);
                 return;
             }
+            // if all presets are present, move on
             if (Arrays.stream(presets).noneMatch(Objects::isNull)) {
+                // check if presets were generated previously and if yes get the EP version
+                try (Stream<String> fileStream = Files.lines(EP_VERSION_PATH)) {
+                    Optional<String> opt = fileStream.findFirst();
+                    if (opt.isPresent()) epVersion = SemanticVersion.parse(opt.get());
+                    ClientMain.LOGGER.info("Found existing presets for EP {}", epVersion.getFriendlyString());
+                } catch (IOException | VersionParsingException e) {
+                    ClientMain.LOGGER.info("No existing presets found");
+                }
+                // scan shader directory for EP shader files and compare versions
+                File targetShader = null;
                 try (Stream<Path> fileStream = Files.walk(SHADER_DIR, 1)) {
-                    fileStream.map(Path::toFile).forEach(shader -> {
+                    Set<File> files = fileStream.map(Path::toFile).collect(Collectors.toUnmodifiableSet());
+                    for (File shader : files) {
+                        String target = "EuphoriaPatches_";
                         String name = shader.getName();
-                        if (name.contains("EuphoriaPatches")) {
-                            for (int i = 0; i < presets.length; i++) {
-                                String preset = presets[i];
-                                if (Arrays.stream(shaderDir.listFiles()).noneMatch(f -> f.getName().equals("Affogato_" + preset + ".zip.txt"))) {
-                                    File shaderCopy = SHADER_DIR.resolve("Affogato_" + preset + ".zip").toFile();
-                                    try {
-                                        if (shader.isDirectory()) FileUtils.copyDirectory(shader, shaderCopy);
-                                        else FileUtils.copyFile(shader, shaderCopy);
-                                        FileUtils.copyFile(
-                                                SHADER_PRESETS_DIR.resolve(i + "_" + preset + ".txt").toFile(),
-                                                SHADER_DIR.resolve("Affogato_" + preset + ".zip.txt").toFile()
-                                        );
-                                        ClientMain.LOGGER.info("Created shader preset {}", preset);
-                                    } catch (IOException e) {
-                                        ClientMain.LOGGER.error("Couldn't copy preset {}, skipping", preset, e);
-                                    }
+                        if (name.contains(target)) {
+                            try {
+                                int beginIndex = name.lastIndexOf(target) + target.length();
+                                SemanticVersion version = shader.isDirectory() ?
+                                        SemanticVersion.parse(name.substring(beginIndex)) :
+                                        SemanticVersion.parse(name.substring(beginIndex, name.lastIndexOf(".")));
+                                if (epVersion == null || version.compareTo((Version) epVersion) > 0) {
+                                    epVersion = version;
+                                    targetShader = shader;
                                 }
+                            } catch (VersionParsingException e) {
+                                ClientMain.LOGGER.info("Couldn't parse version for {}, skipping", name);
                             }
                         }
-                    });
+                    }
                 } catch (IOException e) {
                     ClientMain.LOGGER.error("Can't access file, did not create shader presets", e);
+                }
+                // if no existing presets are found or a newer EP version is found, update shader presets
+                if (epVersion != null && targetShader != null) {
+                    ClientMain.LOGGER.info("Updating Affogato shader presets to EP {}", epVersion.getFriendlyString());
+                    for (int i = 0; i < presets.length; i++) {
+                        String preset = presets[i];
+                        File shaderCopy = SHADER_DIR.resolve("Affogato_" + preset + ".zip").toFile();
+                        try {
+                            if (shaderCopy.exists() && shaderCopy.isDirectory()) FileUtils.deleteDirectory(shaderCopy);
+                            if (targetShader.isDirectory()) FileUtils.copyDirectory(targetShader, shaderCopy);
+                            else FileUtils.copyFile(targetShader, shaderCopy);
+                            FileUtils.copyFile(
+                                    SHADER_PRESETS_DIR.resolve(i + "_" + preset + ".txt").toFile(),
+                                    SHADER_DIR.resolve("Affogato_" + preset + ".zip.txt").toFile()
+                            );
+                            ClientMain.LOGGER.info("Created shader preset {}", preset);
+                        } catch (IOException e) {
+                            ClientMain.LOGGER.error("Couldn't copy preset {}, skipping", preset, e);
+                        }
+                    }
                 }
             } else {
                 ClientMain.LOGGER.warn("Shader preset file name indices not in order, did not create shader presets");
             }
-            if (presets != null && Arrays.stream(presets).noneMatch(Objects::isNull) && Arrays.stream(presets).allMatch(preset -> Arrays.stream(shaderDir.listFiles()).anyMatch(file -> file.getName().equals("Affogato_" + preset + ".zip.txt")))) {
+            // check if all presets got created
+            if (
+                    presets != null && Arrays.stream(presets).noneMatch(Objects::isNull) &&
+                    Arrays.stream(presets).allMatch(preset -> Arrays.stream(shaderDir.listFiles()).anyMatch(file -> file.getName().equals("Affogato_" + preset + ".zip.txt")))
+            ) {
+                // let the Sodium config option know that shader presets can be used
                 ClientMain.LOGGER.info("Shader presets ready");
                 allPresetsCreated = true;
+                // write Affogato EP preset version info to file
+                try (BufferedWriter writer = new BufferedWriter(new FileWriter(EP_VERSION_PATH.toFile()))) {
+                    writer.write(epVersion.getFriendlyString());
+                } catch (IOException e) {
+                    ClientMain.LOGGER.error("Couldn't write version info (how), skipping", e);
+                }
             }
         }
     }